refactor: split up tests for views into separate files
